lib: prevent warning when initializing some bindgen-erated structs
These structs may or may not have padding defined for them on certain
platforms. The build will fail on platforms which do have the padding,
so we need to default it. But, on platforms which do not have the
padding, clippy will complain that we're trying to Default inialize
when we've already intialized all members in the struct. Adding an
`allow` removes the clippy warning.